requirements.txt
PyYAML~=6.0
gym~=0.21.0
numpy~=1.23.2
control~=0.9.4
bagpy~=0.5
pandas~=1.4.4
matplotlib~=3.9.1
setuptools~=49.2.1
scipy~=1.9.1
networkx~=2.8.6
control
tqdm~=4.66.4